requirements.txt
Flask~=2.0.2
PyMySQL~=1.0.2
redis~=4.0.2
DBUtils~=3.0.2
flask-login
flask_cors
flask_sqlalchemy
pyjwt
flask-restful==0.3.9
flasgger==0.9.5